Interface: return appropriate status for bridge device
In function udevInterfaceIsActive, The current design relies on the sys
attribute 'operstate' to determine whether the interface is active.

For the device node of virtual network(say virbr0), There is already a
dummy tap device to maintain a fixed mac on it, So it's available and
its status should be considered as active. But if no anyelse tap device
connects to it, The value of operstate of virbr0 in sysfs is 'down', So
udevInterfaceIsActive returns 0, It causes 'virsh iface-list --all' to
report the status of virbr0 as 'inactive'.

This patch fixes the issue by counting slave devices for bridge device.

Signed-off-by: Lin Ma <lma@suse.com>
Message-Id: <20170213100651.2157-1-lma@suse.com>