qemu: Use full CPU model expansion on x86
The static CPU model expansion is designed to return only canonical
names of all CPU properties. TO maintain backward compatibility libvirt
is stuck with different spelling of some of the features, which is only
returned by the full expansion. But in addition to returned all spelling
variants for all properties the full expansion will contain properties
which are not guaranteed to be migration compatible. We need to combine
both expansions. First we need to call the static expansion to limit the
result to migratable properties. Then we can use the result of the
static expansion as an input to the full expansion to get both canonical
names and their aliases.

Signed-off-by: Jiri Denemark <jdenemar@redhat.com>
Message-Id: <2a96a379f11494760a7bb9026dd2a27a0ee0e3f2.1487176962.git.jdenemar@redhat.com>