conf: Update XML parser, formatter, and RNG schema to support mdev
To keep the domain XML as much platform agnostic as possible, do not
expose an element/attribute which would contain path directly to the
syfs filesystem which the mediated devices are build upon. Instead,
identify each mediated device by a UUID attribute as well as specifying
the device API (e.g. vfio-pci) through the 'model' attribute.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <4465494e61832a8a0e55ef2cf1ba3869d84decb0.1487193773.git.eskultet@redhat.com>