util: Alter virNodeDevice{Create|Delete}Vport to use nodedev APIs
If we have a connection pointer there's no sense walking through the
sysfs in order to create/destroy the vHBA. Instead, let's make use of
the node device create/destroy API's.

Since we don't have to rewrite all the various parent options for
the test driver in order to test whether the storage pool creation
works as the node device creation has been tested already, let's just
use the altered API to test the storage pool paths.

Fix a "bug" in the storage pool test driver code which "assumed"
testStoragePoolObjSetDefaults should fill in the configFile for
both the Define/Create (persistent) and CreateXML (transient) pools
by just VIR_FREE() of the pool during CreateXML.  Because the
configFile was filled in, during Destroy, the pool wouldn't be
free causing a test using the same name pool to fail.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170220131819.2919-7-jferlan@redhat.com>