conf: Add vHBA controller definition to domain
Add the parsing of a vHBA controller to a domain <controller> XML.
The concept of a vHBA is somewhere between a scsi_host <hostdev> and
<controller>, so rather than invent a new <vhba> type, use the
existing <controller> syntax and add an <adapter> element to mimic
the existing storage pool 'fc_host' <adapter> element.

The new element can be parsed, but is unusable. The syntax is:

    <adapter [parent='%s' |
              parent_wwnn='%s' parent_wwpn='%s' |
              parent_fabric_wwn='%s']
              wwnn='%s' wwpn='%s'/>

and is be a subelement of a <controller type='scsi' model='virtio-scsi'>
element. A vHBA is only supported for the virtio-scsi. The controller
will be used exclusively to pass through vHBA LUNs to the domain as they
are discovered or remove them from the domain as that's discovered.

Since this controller is a "virtual" device and no physical qemu -device
is created for it, add checks into various places that would place a disk
or hostdev LUN onto a controller to ensure the vHBA controller is not used.

For now modify the qemu_command and qemu_hotplug code to error out if
one of these is defined. Future patches will add support.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170220131819.2919-9-jferlan@redhat.com>