conf: introduce domain XML element <polling> for iothread
QEMU 2.9.0 will introduce polling feature for AioContext that instead
of blocking syscalls polls for events without blocking.  This means
that polling can be in most cases faster but it also increases CPU
utilization.

To address this issue QEMU implements self-tuning algorithm that
modifies the current polling time to adapt to different workloads
and it can also fallback to blocking syscalls.

For each IOThread this all is controlled by three parameters,
poll-max-ns, poll-grow and poll-shrink.  If parameter poll-max-ns
is set to 0 it disables the polling, if it is omitted the default
behavior is used and any value more than 0 enables polling.
Parameters poll-grow and poll-shrink configure how the self-tuning
algorithm will adapt the current polling time.  If they are omitted
or set to 0 default values will be used.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
Message-Id: <2c15c7d45e334334fb173ad13866b63ee470e812.1487678590.git.phrdina@redhat.com>