nodedev: Add driver callback mechanism to add/remove devices
Add a callback mechanism as a side door of sorts to the event
mgmt functions that are triggered when a node_device object is
added or removed. This includes a mechanism to enumerate already
added devices for those stateInitialize functions called after
the initial nodedevRegister is called. If there is no enumeration
callback, a failure would be returned since the caller would
seem to need that.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170221163328.7070-2-jferlan@redhat.com>