lib: Introduce event for tracking disk backing file write threshold
When using thin provisioning, management tools need to resize the disk
in certain cases. To avoid having them to poll disk fillng introduce an
event whic will be fired when a given offset of the storage is written
by the hypervisor. Together with the API which will be added later, it
will allow to register thresholds for given storage backing volumes and
this event will then notify management if the threshold is exceeded.
Message-Id: <bb267988f2e799afccb39594cf8d9c885047f810.1487877253.git.pkrempa@redhat.com>