qemu_migration: don't allow ABI changes for persistent migration
So far there is probably no change that is allowed to be done
by the VIR_DOMAIN_DEF_PARSE_ABI_UPDATE flag that would break
guest ABI but this may change in the future.

The other cases where this flag is used is only when we are
defining new domain or adding new device into a domain.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
Message-Id: <c7df3c93ad0195786cc965223779e33f3c47f2ab.1488907493.git.phrdina@redhat.com>