qemu: Enforce ACPI, UEFI requirements
Depending on the architecture, requirements for ACPI and UEFI can
be different; more specifically, while on x86 UEFI requires ACPI,
on aarch64 it's the other way around.

Enforce these requirements when building the QEMU command line,
and make the error message more accurate by mentioning that the
requirements are not necessarily applicable to all architectures.

Several aarch64 test cases had to be tweaked because they didn't
respect the requirements.
Message-Id: <1489081057-12677-4-git-send-email-abologna@redhat.com>