qemu: Adaptive timeout for connecting to monitor
There were couple of reports on the list (e.g. [1]) that guests
with huge amounts of RAM are unable to start because libvirt
kills qemu in the initialization phase. The problem is that if
guest is configured to use hugepages kernel has to zero them all
out before handing over to qemu process. For instance, 402GiB
worth of 1GiB pages took around 105 seconds (~3.8GiB/s). Since we
do not want to make the timeout for connecting to monitor
configurable [2], we have to teach libvirt to count with this
fact. This commit implements "1s per each 1GiB of RAM" approach
as suggested here [3].

1: https://www.redhat.com/archives/libvir-list/2017-March/msg00373.html
3: https://www.redhat.com/archives/libvir-list/2017-March/msg00405.html

2: The reason is that ideally someday it will be Libvirt who
creates the monitor socket and qemu will just use it.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <70108c0f23c3e7a35f16bca1a50d6b1933382930.1489405375.git.mprivozn@redhat.com>