conf: Introduce new hostdev device type mdev
A mediated device will be identified by a UUID (with 'model' now being
a mandatory <hostdev> attribute to represent the mediated device API) of
the user pre-created mediated device. The data necessary to identify a
mediated device can be easily extended in the future, e.g. when
auto-creation of mediated devices should be enabled.
Additionally, this patch introduces a new virmdev module under src/util.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <8f901c3db3e198aabe0dcb8aa033c289abb56b70.1489666078.git.eskultet@redhat.com>